AVIServer/src/json_to_postgres/requirements.txt
protobuf==2.6.1
paho-mqtt==1.0
python-dateutil==2.3
psycopg2==2.5.4
pytz==2014.4
argparse==1.4.0
